+++ lib/src/components/supa_email_auth.dart
@@ -387,6 +387,9 @@ class _SupaEmailAuthState extends State<SupaEmailAuth> {
                           heightFactor: 1.0,
                           child: IconButton(
                             iconSize: 20,
+                            tooltip: _isPasswordVisible
+                                ? localization.hidePassword
+                                : localization.showPassword,
                             icon: Icon(_isPasswordVisible
                                 ? Icons.visibility_off_outlined
                                 : Icons.visibility_outlined),
@@ -418,6 +421,9 @@ class _SupaEmailAuthState extends State<SupaEmailAuth> {
                             heightFactor: 1.0,
                             child: IconButton(
                               iconSize: 20,
+                              tooltip: _isConfirmPasswordVisible
+                                  ? localization.hidePassword
+                                  : localization.showPassword,
                               icon: Icon(_isConfirmPasswordVisible
                                   ? Icons.visibility_off_outlined
                                   : Icons.visibility_outlined),