+++ lib/src/components/supa_phone_auth.dart
@@ -100,6 +100,9 @@ class _SupaPhoneAuthState extends State<SupaPhoneAuth> {
                         heightFactor: 1.0,
                         child: IconButton(
                           iconSize: 20,
+                          tooltip: _isPasswordVisible
+                              ? localization.hidePassword
+                              : localization.showPassword,
                           icon: Icon(_isPasswordVisible
                               ? Icons.visibility_off_outlined
                               : Icons.visibility_outlined),